requirements.txt
google-api-python-client==1.9.3
google-cloud-bigquery==1.24.0
pandas==1.0.5
pyarrow==1.0.0
python-igraph==0.8.2
scipy==1.5.2